Added functions to handle shell commands properly.
casapy-friendly shell commands (ls, pwd, etc.) can be called directly from the
casapy prompt, but not from inside a python script.  I added a function to put
these shell commands inside a call to os.system.

os.system can be used to call interactive programs. This is a problem in
benchmarking mode, when interactivity is not desired.  I added a function to
replace 'more' with 'cat' inside the os.system call.  These replacements will
have to be handled on a case-by-case basis.