Benchmarking changes: Exclude raw_input calls; properly suppress tasks.
Exclude raw_input calls from benchmarking scripts.

Replace suppressed tasks with pass statements. This is necessary when a loop
consists entirely of suppressed tasks; otherwise the loop is empty and Python
throws an exception.  I also tacked the suppressed task call onto the end of
the line as a comment.  For example: "pass # plotms(...)". This way we can
clearly see what is being suppressed.