Reworked benchmark scripting code.
benchmark.bash is now the primary script that is called for benchmark
processing.  This script accepts as input a parameters file, specified on the
command line, that tells the script where to acquire the data and the two
casapy scripts it will use for data reduction.

Modified extractCASAscript.py to handle requests for extracting Python from web
pages and requests for benchmarking Python from web pages and Python scripts on
the local network.

Added benchmarking parameters file for data set 2011.0.00367.S.