Explicitly start CASA v4.0.0.
For some unknown reason, the multivac nodes began starting casa 3.4.0 when
casapy was invoked.  This was corrected the following day, but it doesn't
hurt to explicitly start the version I want.

On the down side, this version will have to be updated with the next CASA
release.  I'm leaving it in for now...