Gave user responsibility for execution path, Python path, and analysisUtils.
Added function to check that casa_call module exists in the Python path.

Added README instructions to include distribution directory in the shell
execution path and the in the casapy Python path.

Removed addition of analysisUtils directory from Python module search path.
(It was only available to computers on the NRAO CV network; was not available
from NRAO CV Mac, kwaltz). Since computers not on the CV network will need a
special installation, and since not all CASA guides require analysisUtils, I
will leave it to the user to install and setup analysisUtils when it is needed.
Instructions, referring the user to the analysisUtils wiki page have been
included in the README.